push client to leaf
